backend/requirements.txt
# ClaimIQ Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
httpx==0.27.0

# Supabase
supabase==2.9.0

# ML - Damage Detection
ultralytics==8.3.0
Pillow==10.4.0
numpy==1.26.4

# ML - CLIP (for fraud detection)
# torch and torchvision - install separately based on your system:
# CPU: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# GPU: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# clip: pip install git+https://github.com/openai/CLIP.git

# PDF Report Generation
reportlab==4.2.0

# Utilities
python-dotenv==1.0.1
email-validator==2.1.0
